requirements/base.txt
django==4.0.1
django-widget-tweaks==1.4.12
django-taggit==2.1.0
markdown==3.3.6
python-decouple==3.6
sqlparse
asgiref
gunicorn==20.1.0
psycopg2-binary==2.9.2
django-storages[google]==1.12.3
django-environ==0.8.1
django-debug-toolbar